Move `images` config to separate file.
