+++ .eleventy.js
@@ -2,27 +2,11 @@ const { DateTime } = require("luxon");
 // const pluginRss = require("@11ty/eleventy-plugin-rss");
 const pluginSyntaxHighlight = require("@11ty/eleventy-plugin-syntaxhighlight");
 const eleventyNavigationPlugin = require("@11ty/eleventy-navigation");
-const Image = require("@11ty/eleventy-img");
-const svgContents = require("eleventy-plugin-svg-contents");
-
-async function imageShortcode(src, alt, sizes) {
-  let metadata = await Image(src, {
-    widths: [150, 300, 600],
-    formats: ["avif", "jpeg"],
-    outputDir: "./public/img/",
-  });
-
-  let imageAttributes = {
-    alt,
-    sizes,
-    loading: "lazy",
-    decoding: "async",
-  };
-  return Image.generateHTML(metadata, imageAttributes);
-}
 
 module.exports = function (eleventyConfig) {
-  eleventyConfig.addWatchTarget("./src/sass/");
+	// add Image config
+	eleventyConfig.addPlugin(require("./config/eleventy.config.images.js"));
+	eleventyConfig.addWatchTarget("./src/sass/");
   eleventyConfig.addPassthroughCopy("./src/css");
 	eleventyConfig.addPassthroughCopy({"./src/assets/favicons": "/favicons"});
 	eleventyConfig.addPassthroughCopy({
@@ -31,26 +15,12 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.addPlugin(eleventyNavigationPlugin);
 
   // eleventyConfig.addPlugin(pluginRss);
-	eleventyConfig.addPlugin(svgContents);
   eleventyConfig.addPlugin(pluginSyntaxHighlight);
-	eleventyConfig.addNunjucksAsyncShortcode("image", imageShortcode);
-  eleventyConfig.addLiquidShortcode("image", imageShortcode);
-  eleventyConfig.addJavaScriptFunction("image", imageShortcode);
-
   eleventyConfig.addFilter("readableDate", (dateObj) => {
     return DateTime.fromJSDate(dateObj, { zone: "utc" }).toFormat(
       "dd LLL yyyy"
     );
   });
-	// SVGs
-  eleventyConfig.addNunjucksAsyncShortcode("svgIcon", async (filename) => {
-    const metadata = await Image(`./src/assets/icons/${filename}`, {
-      formats: ["svg"],
-      dryRun: true,
-    });
-    return metadata.svg[0].buffer.toString();
-  });
-
   // https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#valid-date-string
   eleventyConfig.addFilter("htmlDateString", (dateObj) => {
     return DateTime.fromJSDate(dateObj, { zone: "utc" }).toFormat("yyyy-LL-dd");
