+++ config/eleventy.config.images.js
@@ -1,53 +1,34 @@
-const path = require("path");
-const eleventyImage = require("@11ty/eleventy-img");
+const Image = require("@11ty/eleventy-img");
+const svgContents = require("eleventy-plugin-svg-contents");
 
-function relativeToInputPath(inputPath, relativeFilePath) {
-  let split = inputPath.split("/");
-  split.pop();
 
-  return path.resolve(split.join(path.sep), relativeFilePath);
-}
+async function imageShortcode(src, alt, sizes) {
+  let metadata = await Image(src, {
+    widths: [150, 300, 600],
+    formats: ["avif", "jpeg"],
+    outputDir: "./public/img/",
+  });
 
-function isFullUrl(url) {
-  try {
-    new URL(url);
-    return true;
-  } catch (e) {
-    return false;
-  }
+  let imageAttributes = {
+    alt,
+    sizes,
+    loading: "lazy",
+    decoding: "async",
+  };
+  return Image.generateHTML(metadata, imageAttributes);
 }
 
 module.exports = function (eleventyConfig) {
-  // Eleventy Image shortcode
-  // https://www.11ty.dev/docs/plugins/image/
-  eleventyConfig.addAsyncShortcode(
-    "image",
-    async function imageShortcode(src, alt, widths, sizes) {
-      // Full list of formats here: https://www.11ty.dev/docs/plugins/image/#output-formats
-      // Warning: Avif can be resource-intensive so take care!
-      let formats = ["avif", "webp", "auto"];
-      let input;
-      if (isFullUrl(src)) {
-        input = src;
-      } else {
-        input = relativeToInputPath(this.page.inputPath, src);
-      }
-
-      let metadata = await eleventyImage(input, {
-        widths: widths || ["auto"],
-        formats,
-        outputDir: path.join(eleventyConfig.dir.output, "img"), // Advanced usage note: `eleventyConfig.dir` works here because we’re using addPlugin.
-      });
-
-      // TODO loading=eager and fetchpriority=high
-      let imageAttributes = {
-        alt,
-        sizes,
-        loading: "lazy",
-        decoding: "async",
-      };
-
-      return eleventyImage.generateHTML(metadata, imageAttributes);
-    }
-  );
+  eleventyConfig.addPlugin(svgContents);
+  eleventyConfig.addNunjucksAsyncShortcode("image", imageShortcode);
+  eleventyConfig.addLiquidShortcode("image", imageShortcode);
+  eleventyConfig.addJavaScriptFunction("image", imageShortcode);
+  // SVGs
+  eleventyConfig.addNunjucksAsyncShortcode("svgIcon", async (filename) => {
+    const metadata = await Image(`./src/assets/icons/${filename}`, {
+      formats: ["svg"],
+      dryRun: true,
+    });
+    return metadata.svg[0].buffer.toString();
+  });
 };
